# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ember local-first productivity API — core loop for CLI and MCP clients.
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export type DragonType = (typeof DragonType)[keyof typeof DragonType];

export const DragonType = {
  cinder: "cinder",
  moss: "moss",
  drift: "drift",
  frost: "frost",
} as const;

export type DragonStage = (typeof DragonStage)[keyof typeof DragonStage];

export const DragonStage = {
  egg: "egg",
  hatchling: "hatchling",
  adolescent: "adolescent",
  adult: "adult",
  ancient: "ancient",
} as const;

export interface Project {
  id: string;
  name: string;
  dragon_type: DragonType;
  dragon_stage: DragonStage;
  total_focus_minutes: number;
  project_summary: string;
  created_at: string;
  updated_at: string;
  last_session_at?: string | null;
  last_decay_check?: string | null;
  is_archived: number;
}

export interface CreateProjectRequest {
  name: string;
  dragon_type: DragonType;
  summary?: string;
}

export interface UpdateProjectRequest {
  name?: string;
  project_summary?: string;
}

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  active: "active",
  backlog: "backlog",
  completed: "completed",
} as const;

export type TaskSource = (typeof TaskSource)[keyof typeof TaskSource];

export const TaskSource = {
  ai: "ai",
  user: "user",
  reflection: "reflection",
} as const;

export interface Task {
  id: string;
  project_id: string;
  task_text: string;
  status: TaskStatus;
  priority: number;
  task_order: number;
  source: TaskSource;
  created_at: string;
  completed_at?: string | null;
}

export interface CreateTaskRequest {
  project_id: string;
  task_text: string;
  source?: TaskSource;
  status?: TaskStatus;
}

export type UpdateTaskRequestAction =
  (typeof UpdateTaskRequestAction)[keyof typeof UpdateTaskRequestAction];

export const UpdateTaskRequestAction = {
  complete: "complete",
  move_to_active: "move_to_active",
  move_to_backlog: "move_to_backlog",
} as const;

export interface UpdateTaskRequest {
  action?: UpdateTaskRequestAction;
  task_text?: string;
  status?: TaskStatus;
}

export interface Session {
  id: string;
  project_id: string;
  start_time: string;
  end_time?: string | null;
  duration_minutes: number;
  reflection?: string | null;
  ai_summary?: string | null;
  tasks_completed_count: number;
  created_at: string;
}

export interface StartSessionRequest {
  project_id: string;
  task_ids?: string[];
}

export interface EndSessionRequest {
  session_id: string;
  reflection?: string;
  tasks_completed_count?: number;
}

export interface EndSessionResponse {
  session?: Session;
  project?: Project;
  previous_dragon_stage?: DragonStage | null;
}

export type ResumeResponseSource =
  (typeof ResumeResponseSource)[keyof typeof ResumeResponseSource];

export const ResumeResponseSource = {
  ai: "ai",
  fallback: "fallback",
} as const;

export interface ResumeResponse {
  status_summary: string;
  suggested_next_step: string;
  last_session_summary?: string | null;
  source: ResumeResponseSource;
}

export interface AiStatus {
  available: boolean;
}

export interface ExtractTasksRequest {
  project_id: string;
  user_input: string;
}

export interface ProcessReflectionRequest {
  project_id: string;
  session_id: string;
  reflection: string;
}

export interface AiExtractionResult {
  new_active_tasks?: string[];
  new_backlog_tasks?: string[];
  completed_tasks?: string[];
  insights?: string[];
  blockers?: string[];
  summary_update?: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface ErrorResponse {
  error: string;
}

export type TrustBand = (typeof TrustBand)[keyof typeof TrustBand];

export const TrustBand = {
  paired: "paired",
  solo: "solo",
  autonomous: "autonomous",
} as const;

export type SkillMode = (typeof SkillMode)[keyof typeof SkillMode];

export const SkillMode = {
  paired: "paired",
  autonomous: "autonomous",
} as const;

export type SkillRunStatus =
  (typeof SkillRunStatus)[keyof typeof SkillRunStatus];

export const SkillRunStatus = {
  pending: "pending",
  approved: "approved",
  edited: "edited",
  rejected: "rejected",
  failed: "failed",
  over_budget: "over_budget",
} as const;

export type Verdict = (typeof Verdict)[keyof typeof Verdict];

export const Verdict = {
  approve: "approve",
  edit: "edit",
  reject: "reject",
} as const;

export interface Skill {
  id: string;
  name: string;
  description: string;
  agent_recipe_key: string;
  default_trust_band: TrustBand;
  cost_estimate_input_tokens: number;
  created_at: string;
}

export interface DragonSkillMaturity {
  dragon_id?: string;
  skill_id?: string;
  runs?: number;
  approvals?: number;
  edits?: number;
  rejections?: number;
  current_trust?: TrustBand;
  locked_band?: TrustBand | null;
  paused?: number;
  last_used_at?: string | null;
  created_at?: string;
}

export interface DragonSkillEntry {
  skill: Skill;
  maturity: DragonSkillMaturity;
}

export type SkillRunComplexity =
  (typeof SkillRunComplexity)[keyof typeof SkillRunComplexity];

export const SkillRunComplexity = {
  simple: "simple",
  complex: "complex",
} as const;

export interface SkillRun {
  id: string;
  dragon_id: string;
  skill_id: string;
  project_id: string;
  mode: SkillMode;
  complexity?: SkillRunComplexity;
  user_prompt: string;
  system_prompt?: string;
  output_text?: string | null;
  user_edit?: string | null;
  status: SkillRunStatus;
  model?: string;
  input_tokens?: number;
  output_tokens?: number;
  cost_usd?: number;
  ran_at: string;
  verdicted_at?: string | null;
}

export interface InvokeSkillRequest {
  user_prompt: string;
  mode?: SkillMode;
  confirm_high_cost?: boolean;
}

export interface InvokeSkillResponse {
  ok?: boolean;
  run?: SkillRun;
  error?: string;
  required_trust?: TrustBand;
  current_trust?: TrustBand;
  estimated_cost_usd?: number;
}

export type SkillVerdictRequestStatus =
  (typeof SkillVerdictRequestStatus)[keyof typeof SkillVerdictRequestStatus];

export const SkillVerdictRequestStatus = {
  approved: "approved",
  edited: "edited",
  rejected: "rejected",
} as const;

export interface SkillVerdictRequest {
  verdict?: Verdict;
  status?: SkillVerdictRequestStatus;
  user_edit?: string;
}

export interface SkillVerdictResponse {
  ok?: boolean;
  run?: SkillRun;
  error?: string;
}

export interface InboxItem {
  run?: SkillRun;
  skill_name?: string;
  pause_state?: boolean;
}

export type InboxResponsePausedSkillsItem = {
  skill_id?: string;
  skill_name?: string;
};

export interface InboxResponse {
  pending?: InboxItem[];
  recently_handled?: InboxItem[];
  paused_skills?: InboxResponsePausedSkillsItem[];
  high_cost_pending?: InboxItem[];
}

/**
 * Not found
 */
export type NotFoundResponse = ErrorResponse;

export type ListProjectsParams = {
  archived?: boolean;
};

export type ListTasksParams = {
  project_id: string;
  status?: TaskStatus;
};

export type ListProjectSessionsParams = {
  limit?: number;
};

export type GetResumeParams = {
  project_id: string;
};

export type ListDragonSkillRunsParams = {
  limit?: number;
};