# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Ember local-first productivity API — core loop for CLI and MCP clients.
servers:
  - url: /api
    description: Base API path (proxied from frontend; default http://localhost:8080/api)
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects
      parameters:
        - name: archived
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Hatch a new dragon project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by id
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project fields
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: archiveProject
      tags: [projects]
      summary: Archive project
      parameters:
        - $ref: "#/components/parameters/ProjectId"
      responses:
        "200":
          description: Archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /ai/extract-tasks:
    post:
      operationId: extractTasks
      tags: [ai]
      summary: Brain dump — extract tasks from unstructured text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtractTasksRequest"
      responses:
        "200":
          description: Extraction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiExtractionResult"
        "503":
          description: AI unavailable
  /ai/process-reflection:
    post:
      operationId: processReflection
      tags: [ai]
      summary: Process session reflection via AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessReflectionRequest"
      responses:
        "200":
          description: Reflection processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiExtractionResult"
...
components:
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    DragonType:
      type: string
      enum: [cinder, moss, drift, frost]
    DragonStage:
      type: string
      enum: [egg, hatchling, adolescent, adult, ancient]
    Project:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        dragon_type: { $ref: "#/components/schemas/DragonType" }
        dragon_stage: { $ref: "#/components/schemas/DragonStage" }
        total_focus_minutes: { type: integer }
        project_summary: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        last_session_at: { type: string, format: date-time, nullable: true }
        last_decay_check: { type: string, format: date-time, nullable: true }
        is_archived: { type: integer }
      required:
        - id
        - name
        - dragon_type
        - dragon_stage
        - total_focus_minutes
        - project_summary
        - created_at
        - updated_at
        - is_archived
    CreateProjectRequest:
      type: object
      properties:
        name: { type: string }
        dragon_type: { $ref: "#/components/schemas/DragonType" }
        summary: { type: string }
      required: [name, dragon_type]
    UpdateProjectRequest:
      type: object
      properties:
        name: { type: string }
        project_summary: { type: string }
...
    ExtractTasksRequest:
      type: object
      properties:
        project_id: { type: string }
        user_input: { type: string }
      required: [project_id, user_input]
    ProcessReflectionRequest:
      type: object
      properties:
        project_id: { type: string }
        session_id: { type: string }
        reflection: { type: string }
      required: [project_id, session_id, reflection]
    AiExtractionResult:
      type: object
      properties:
        new_active_tasks:
          type: array
          items: { type: string }
        new_backlog_tasks:
          type: array
          items: { type: string }
        completed_tasks:
          type: array
          items: { type: string }
        insights:
          type: array
          items: { type: string }
        blockers:
          type: array
          items: { type: string }
        summary_update: { type: string }
    SuccessResponse:
      type: object
      properties:
        success: { type: boolean }
      required: [success]
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
...